requirements.txt
streamlit
tensorflow-cpu
opencv-python-headless
numpy
Pillow